+++ README.md
@@ -16,7 +16,8 @@
 ## Step2 진행 중 의식의 흐름
 
 - Parcelable이 기본 세팅 되어 있지 않은 점으로 미루어 볼 때, 이번 미션에서는 사용하지 않는 것이 좋을 것 같다고 판단했습니다. 컴포즈 학습에 집중하기 위해서요.
-- 지난 회원가입 미션에서는 컴포저블을 최대한 재활용 해보려고 애쓰다가 불필요한 로직이 추가되었습니다. 이번 미션의 이번 단계에서는 일단 최대한 각 화면을 분리하여 구현해 보려고 합니다.
+- 지난 회원가입 미션에서는 컴포저블을 최대한 재활용 해보려고 애쓰다가 불필요한 로직이 추가되었습니다. 이번 미션의 이번 단계에서는 일단 최대한 각 화면을 분리하여 구현해 보려고
+  합니다.
 
 ## Step1 구현 목록
