+++ app/src/androidTest/java/nextstep/shoppingcart/view/ProductItemTest.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import org.junit.Rule
 import org.junit.Test
 import nextstep.shoppingcart.model.Product
+import nextstep.shoppingcart.view.product.ProductItem
 import org.junit.Before
 
 class ProductItemTest {
