+++ app/src/androidTest/java/nextstep/shoppingcart/view/ProductsGridTest.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import org.junit.Rule
 import org.junit.Test
 import nextstep.shoppingcart.model.Product
+import nextstep.shoppingcart.view.product.ProductsGrid
 import org.junit.Before
 
 class ProductsGridTest {
