+++ app/src/androidTest/java/nextstep/shoppingcart/view/ProductsTopAppBarTest.kt
@@ -3,6 +3,7 @@ package nextstep.shoppingcart.view
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
+import nextstep.shoppingcart.view.product.ProductsTopAppBar
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
