+++ app/src/main/AndroidManifest.xml
@@ -15,14 +15,14 @@
         android:theme="@style/Theme.ShoppingCart"
         tools:targetApi="31">
         <activity
-            android:name=".view.ProductDetailActivity"
+            android:name=".view.product.detail.ProductDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".view.CartActivity"
+            android:name=".view.cart.CartActivity"
             android:exported="false"
             android:theme="@style/Theme.ShoppingCart" />
         <activity
-            android:name=".view.ProductsActivity"
+            android:name=".view.product.ProductsActivity"
             android:exported="true"
             android:theme="@style/Theme.ShoppingCart">
             <intent-filter>
