+++ app/src/main/java/nextstep/shoppingcart/view/product/ProductItem.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.view
+package nextstep.shoppingcart.view.product
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
