+++ app/src/main/java/nextstep/shoppingcart/view/product/ProductsActivity.kt
@@ -1,14 +1,14 @@
-package nextstep.shoppingcart.view
+package nextstep.shoppingcart.view.product
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 
-class ProductDetailActivity : ComponentActivity() {
+class ProductsActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            ProductDetailScreen()
+            ProductsScreen()
         }
     }
 }
